+++ LectureApp/src/main/java/com/hhp/lectureapp/lecture/business/dto/PostLectureDto.java
@@ -1,4 +1,4 @@
-package com.hhp.lectureapp.lecture.controller.dto;
+package com.hhp.lectureapp.lecture.business.dto;
 
 public class PostLectureDto {
     private int lectureId;
